# internal/sqlc/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.31.1

package db

import (
	"github.com/google/uuid"
	"github.com/jackc/pgx/v5/pgtype"
)

type Challenge struct {
	ID          uuid.UUID          `json:"id"`
	MerchantID  uuid.UUID          `json:"merchant_id"`
	Title       string             `json:"title"`
	Description string             `json:"description"`
	TargetMiles float64            `json:"target_miles"`
	ExpiresAt   pgtype.Timestamptz `json:"expires_at"`
	CreatedAt   pgtype.Timestamptz `json:"created_at"`
}

type ChallengeRegistration struct {
	ID           uuid.UUID          `json:"id"`
	ChallengeID  uuid.UUID          `json:"challenge_id"`
	UserID       uuid.UUID          `json:"user_id"`
	RegisteredAt pgtype.Timestamptz `json:"registered_at"`
	Status       string             `json:"status"`
}

type Employee struct {
	ID         uuid.UUID          `json:"id"`
	MerchantID uuid.UUID          `json:"merchant_id"`
	Name       string             `json:"name"`
	Phone      string             `json:"phone"`
	Code       string             `json:"code"`
	Status     string             `json:"status"`
	CreatedAt  pgtype.Timestamptz `json:"created_at"`
}

type Merchant struct {
	ID          uuid.UUID          `json:"id"`
	OwnerUserID uuid.UUID          `json:"owner_user_id"`
	Name        string             `json:"name"`
	Category    string             `json:"category"`
	AddressLat  float64            `json:"address_lat"`
	AddressLng  float64            `json:"address_lng"`
	LogoUrl     string             `json:"logo_url"`
	Description string             `json:"description"`
	CreatedAt   pgtype.Timestamptz `json:"created_at"`
}

type Otp struct {
	ID        uuid.UUID          `json:"id"`
	Phone     string             `json:"phone"`
	CodeHash  string             `json:"code_hash"`
	ExpiresAt pgtype.Timestamptz `json:"expires_at"`
	UsedAt    pgtype.Timestamptz `json:"used_at"`
	CreatedAt pgtype.Timestamptz `json:"created_at"`
}

type Session struct {
	ID            uuid.UUID          `json:"id"`
	UserID        uuid.UUID          `json:"user_id"`
	ChallengeID   uuid.UUID          `json:"challenge_id"`
	StartTime     pgtype.Timestamptz `json:"start_time"`
	EndTime       pgtype.Timestamptz `json:"end_time"`
	StartLat      float64            `json:"start_lat"`
	StartLng      float64            `json:"start_lng"`
	EndLat        float64            `json:"end_lat"`
	EndLng        float64            `json:"end_lng"`
	StepsStart    int32              `json:"steps_start"`
	StepsEnd      int32              `json:"steps_end"`
	MilesStart    float64            `json:"miles_start"`
	MilesEnd      float64            `json:"miles_end"`
	DistanceMiles float64            `json:"distance_miles"`
	Status        string             `json:"status"`
	HmacSecret    string             `json:"hmac_secret"`
	CreatedAt     pgtype.Timestamptz `json:"created_at"`
	UpdatedAt     pgtype.Timestamptz `json:"updated_at"`
}

type SessionCheckpoint struct {
	ID             int64              `json:"id"`
	SessionID      uuid.UUID          `json:"session_id"`
	Lat            float64            `json:"lat"`
	Lng            float64            `json:"lng"`
	RecordedAt     pgtype.Timestamptz `json:"recorded_at"`
	Steps          int32              `json:"steps"`
	DistanceMeters float64            `json:"distance_meters"`
	SpeedMps       float64            `json:"speed_mps"`
	SpeedViolation bool               `json:"speed_violation"`
}

type User struct {
	ID        uuid.UUID          `json:"id"`
	Phone     string             `json:"phone"`
	Name      string             `json:"name"`
	AvatarUrl string             `json:"avatar_url"`
	Role      string             `json:"role"`
	CreatedAt pgtype.Timestamptz `json:"created_at"`
}

type Voucher struct {
	ID                   uuid.UUID          `json:"id"`
	UserID               uuid.UUID          `json:"user_id"`
	ChallengeID          uuid.UUID          `json:"challenge_id"`
	SessionID            uuid.UUID          `json:"session_id"`
	Code                 string             `json:"code"`
	Status               string             `json:"status"`
	IssuedAt             pgtype.Timestamptz `json:"issued_at"`
	RedeemedAt           pgtype.Timestamptz `json:"redeemed_at"`
	RedeemedByEmployeeID pgtype.UUID        `json:"redeemed_by_employee_id"`
}

type VoucherRedemption struct {
	ID         uuid.UUID          `json:"id"`
	VoucherID  uuid.UUID          `json:"voucher_id"`
	EmployeeID uuid.UUID          `json:"employee_id"`
	RedeemedAt pgtype.Timestamptz `json:"redeemed_at"`
}