# openapi.yaml
openapi: 3.0.3
info:
  title: OMM API
  version: 0.1.0
servers:
  - url: /api
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
    AuthOtpRequest:
      type: object
      required: [phone]
      properties:
        phone:
          type: string
    AuthOtpResponse:
      type: object
      properties:
        otp_id:
          type: string
          format: uuid
        expires_at:
          type: string
          format: date-time
        debug_code:
          type: string
    AuthVerifyRequest:
      type: object
      required: [otp_id, phone, code]
      properties:
        otp_id:
          type: string
          format: uuid
        phone:
          type: string
        code:
          type: string
    AuthVerifyResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        phone:
          type: string
        name:
          type: string
        avatar_url:
          type: string
        role:
          type: string
        created_at:
          type: string
          format: date-time
...
    Challenge:
      type: object
      properties:
        id:
          type: string
          format: uuid
        merchant_id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        target_miles:
          type: number
        expires_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
    CreateChallengeRequest:
      type: object
      required: [title, target_miles, expires_at]
      properties:
        title:
          type: string
        description:
          type: string
        target_miles:
          type: number
        expires_at:
          type: string
          format: date-time
    ChallengeRegistration:
      type: object
      properties:
        id:
          type: string
          format: uuid
        challenge_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        registered_at:
          type: string
          format: date-time
        status:
          type: string
...
    MerchantRegisterRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        category:
          type: string
        address_lat:
          type: number
        address_lng:
          type: number
        logo_url:
          type: string
        description:
          type: string
    Merchant:
      type: object
      properties:
        id:
          type: string
          format: uuid
        owner_user_id:
          type: string
          format: uuid
        name:
          type: string
        category:
          type: string
        address_lat:
          type: number
        address_lng:
          type: number
        logo_url:
          type: string
        description:
          type: string
    MerchantDashboard:
      type: object
      properties:
        total_challenges:
          type: integer
          format: int64
        total_redemptions:
          type: integer
          format: int64
        active_customers:
          type: integer
          format: int64
    MerchantChallenge:
      allOf:
        - $ref: "#/components/schemas/Challenge"
        - type: object
          properties:
            participants:
              type: integer
              format: int64
    CreateEmployeeRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        phone:
          type: string
    Employee:
      type: object
      properties:
        id:
          type: string
          format: uuid
        merchant_id:
          type: string
          format: uuid
        name:
          type: string
        phone:
          type: string
        code:
          type: string
        status:
          type: string
...
paths:
  /auth/otp:
    post:
      summary: Send OTP
      operationId: sendOtp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthOtpRequest"
      responses:
        "200":
          description: OTP created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthOtpResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/verify:
    post:
      summary: Verify OTP
      operationId: verifyOtp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthVerifyRequest"
      responses:
        "200":
          description: Token issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthVerifyResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /challenges:
    get:
      summary: List challenges
      operationId: listChallenges
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Challenge list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Challenge"
    post:
      summary: Create challenge
      operationId: createChallenge
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateChallengeRequest"
      responses:
        "200":
          description: Challenge created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Challenge"
  /challenges/{id}/register:
    post:
      summary: Register for challenge
      operationId: registerChallenge
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Registration created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChallengeRegistration"
  /merchants/register:
    post:
      summary: Register merchant
      operationId: registerMerchant
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MerchantRegisterRequest"
      responses:
        "200":
          description: Merchant created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Merchant"
  /merchants/dashboard:
    get:
      summary: Merchant dashboard
      operationId: merchantDashboard
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MerchantDashboard"
  /merchants/challenges:
    get:
      summary: List merchant challenges
      operationId: listMerchantChallenges
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Merchant challenges
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MerchantChallenge"
  /merchants/employees:
    get:
      summary: List employees
      operationId: listEmployees
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Employee list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Employee"
    post:
      summary: Create employee
      operationId: createEmployee
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEmployeeRequest"
      responses:
        "200":
          description: Employee created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
...